chore: separated nginx less test from official agent testing
